+++ .editorconfig
@@ -39,6 +39,10 @@ indent_style = space
 indent_size = 4
 trim_trailing_whitespace = false
 
+[.circleci/config.yml]
+indent_style = space
+indent_size = 2
+
 [.gitmodules]
 indent_style = tab
