+++ phpunit.xml.dist
@@ -12,7 +12,7 @@
         <ini name="memory_limit" value="-1" />
         <server name="KERNEL_DIR" value="tests/Fixtures/app/" />
         <server name="KERNEL_CLASS" value="AppKernel" />
-        <server name="APP_ENV" value="test_phpunit" />
+        <server name="APP_ENV" value="test" />
         <server name="LEGACY" value="0" />
     </php>
