+++ tests/Bridge/Doctrine/Orm/Filter/ExistsFilterTest.php
@@ -14,12 +14,13 @@ declare(strict_types=1);
 namespace ApiPlatform\Core\Tests\Bridge\Doctrine\Orm\Filter;
 
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\ExistsFilter;
+use ApiPlatform\Core\Test\DoctrineOrmFilterTestCase;
 use ApiPlatform\Core\Tests\Fixtures\TestBundle\Entity\Dummy;
 
 /**
  * @author Antoine Bluchet <soyuka@gmail.com>
  */
-class ExistsFilterTest extends AbstractFilterTest
+class ExistsFilterTest extends DoctrineOrmFilterTestCase
 {
     protected $filterClass = ExistsFilter::class;
