+++ tests/Bridge/Doctrine/Orm/Filter/RangeFilterTest.php
@@ -14,12 +14,13 @@ declare(strict_types=1);
 namespace ApiPlatform\Core\Tests\Bridge\Doctrine\Orm\Filter;
 
 use ApiPlatform\Core\Bridge\Doctrine\Orm\Filter\RangeFilter;
+use ApiPlatform\Core\Test\DoctrineOrmFilterTestCase;
 use ApiPlatform\Core\Tests\Fixtures\TestBundle\Entity\Dummy;
 
 /**
  * @author Lee Siong Chan <ahlee2326@me.com>
  */
-class RangeFilterTest extends AbstractFilterTest
+class RangeFilterTest extends DoctrineOrmFilterTestCase
 {
     protected $filterClass = RangeFilter::class;
