--- tests/Fixtures/app/config/config_test_phpunit.yml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: config_test.yml }
